編集

+++ HelloWorld/src/helloworld/EnzanshiP97.java
@@ -6,8 +6,10 @@ public class EnzanshiP97 {
 		System.out.println("身長は"+dnum+"です。");
 		
 		System.out.println("double型の変数に代入します。");
-		int inum=dnum;
-		//このように、大きい型から小さい型へ代入することはできません。
+		int inum=(int)dnum;
+		//大きい型から小さい型へ代入することはできません。
+		//小さい型へと代入するには　int inum=(int)dnum;(9行目)のように変換する必要があります。
+		//このように (int) の部分を「キャスト演算子」と呼びます。
 		
 		System.out.println("身長は"+inum+"です。");
 	}
